refactor: rename plugin_utils to fsstate
