+++ teal/packer/actions.tl
@@ -5,7 +5,7 @@ local a      = require 'packer.async'
 local config = require 'packer.config'
 local log    = require 'packer.log'
 local util   = require 'packer.util'
-local plugin_utils = require 'packer.plugin_utils'
+local fsstate = require 'packer.fsstate'
 
 local display = require 'packer.display'
 
@@ -211,7 +211,7 @@ local function get_install_tasks(
   return tasks
 end
 
-local function move_plugin(plugin: Plugin, moves: {string:Display.Result}, fs_state: plugin_utils.FSState)
+local function move_plugin(plugin: Plugin, moves: {string:Display.Result}, fs_state: fsstate.FSState)
   local from: string
   local to: string
   if plugin.opt then
@@ -306,7 +306,7 @@ local function fix_plugin_types(
   plugins: {string:Plugin},
   extra_plugins: {string},
   moves: {string:Display.Result},
-  fs_state: plugin_utils.FSState
+  fs_state: fsstate.FSState
 )
   log.debug 'Fixing plugin types'
   -- NOTE: This function can only be run on plugins already installed
@@ -341,8 +341,8 @@ local function filter_opts_from_plugins(first: table|string, ...: string): table
 end
 
 -- Find and remove any plugins not currently configured for use
-local do_clean = a.sync(function(plugins: {string:Plugin}, fs_state: plugin_utils.FSState, removals: {string})
-  fs_state = fs_state or plugin_utils.get_fs_state(plugins)
+local do_clean = a.sync(function(plugins: {string:Plugin}, fs_state: fsstate.FSState, removals: {string})
+  fs_state = fs_state or fsstate.get_fs_state(plugins)
 
   log.debug 'Starting clean'
   local extra_plugins = fs_state.extra
@@ -383,7 +383,7 @@ end, 4)
 M.install = a.sync(function()
   log.debug 'packer.install: requiring modules'
 
-  local fs_state = plugin_utils.get_fs_state(_G.packer_plugins)
+  local fs_state = fsstate.get_fs_state(_G.packer_plugins)
   local missing_plugins = vim.tbl_values(fs_state.missing)
   if #missing_plugins == 0 then
     log.info 'All configured plugins are installed'
@@ -416,7 +416,7 @@ end)
 M.update = a.void(function(first: string|table, ...:string)
   local plugins = _G.packer_plugins
   local opts, update_plugins = filter_opts_from_plugins(first, ...)
-  local fs_state = plugin_utils.get_fs_state(plugins)
+  local fs_state = fsstate.get_fs_state(plugins)
   local _, installed_plugins = util.partition(vim.tbl_values(fs_state.missing), update_plugins)
 
   local updates = {}
@@ -451,7 +451,7 @@ end)
 M.sync = a.void(function(first: string|table, ...:string)
   local plugins = _G.packer_plugins
   local opts, update_plugins = filter_opts_from_plugins(first, ...)
-  local fs_state = plugin_utils.get_fs_state(plugins)
+  local fs_state = fsstate.get_fs_state(plugins)
 
   local extra_plugins = util.partition(vim.tbl_values(fs_state.extra), update_plugins)
 
@@ -466,7 +466,7 @@ M.sync = a.void(function(first: string|table, ...:string)
 
   -- Even though we may have moved some dirty plugins, they may still be dirty
   -- for a different reason so recalculate fs_state
-  fs_state = plugin_utils.get_fs_state(plugins)
+  fs_state = fsstate.get_fs_state(plugins)
 
   do_clean(plugins, fs_state, results.removals)
